requirements-dev.txt
-r requirements.txt
black==21.9b0
coverage==5.5
coveralls==3.2.0
factory-boy==3.2.0
Faker==9.5.2
flake8==4.0.1
isort==5.9.3
pre-commit==2.15.0
pytest-django==4.4.0
responses==0.14.0

# mypy and PEP 561 stubs
mypy==0.910
django-stubs==1.9.0
djangorestframework-stubs==1.4.0
types-requests==2.25.11
